+++ src/main/java/notai/aiTask/domain/AITaskRepository.java
@@ -1,4 +1,4 @@
-package notai.ai_task.domain;
+package notai.aiTask.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
